v0.2
A number of features added and bugs fixed/scenarios handled which were
not in v0.1.

Features:

* More than one type in an individual node now handled
* Handling of OWL metacharacters and spaces in individual IRIs
  now configurable (flexibly) by command-line options. An rdfs:label
annotation property with the original text in an individual node added
by default, can be disabled by a command line option.
* The parser now attempts to give an informative error if one tries to
  add an arrow between literals.

Bugs fixed/scenarios now handled:

* Geometry guesswork added for literal nodes
* Geometry guesswork tweaked to handle (recursively) the case that the
  arrow is part of a group
* Linebreaks within text/HTML, for example within literal nodes, now
  handled in a way which seems about as robust as possible (this was
buggy in v0.1, and the bug obscured the need for robust prettification)
* A bug in which metacharacters should have been handled but were not
  has been fixed
* Various uninformative error messages have been improved to give cell
  IDs.