Add additional example and tests
We add an example created and kindly shared by Aaron Hope. We also add
a test which runs through all of the .drawio files in the 'examples'
directory and checks that the generated OWL without preamble is as
expected; that the generated OWL with preamble is as expected is also
tested for one of the files.