parser v0.2.1-main.sha.1171107+gbad.1: add some obj properties
Some valid RiC-O 1.0.1 object properties
were missing from the in-built script dictionary
but were present in the drawio file, so I added them.
In future versions, it is definitely worth
having the parser link to the actual ontology
and draw its dictionary from there, but that's
for another time.

The versioning is a bit strange because I had
to rebase to main but also want to keep
our own gbad versioning separate.

The drawio file I wanted to parse (and parsed
successfully) will be added in the next commit.