Convert Orville Wood authority using -m url option and fixed GBAD namespaces
The -m url option is defined in the parent commit whereas
the namespaces are used correctly for the first time;
my previous conversion specified the base namespace
in the ontology option, for which reason they were parsed
with a hash, so now this is fixed and the appropriate ontology
is also referenced.

I am not quite sure at the moment, though, if
/schema/authority# and /schema/description-listings#
will eventually be used for respective ontologies
because I will need to make this all work with RML
(with the general schema files); therefore, this part is
subject to change.

I created a bash script for easier conversions that
I will commit in the next commit.