Add a shell script for easier conversions
By default the script uses gbad-specific values, obviously,
but these can be replaced with custom ones; also,
custom parameters can be passed after the file name,
so these will be relayed to the python script.

So, essentially, this is just a wrapper for the python script
that allows to set some default options.