Auto-convert all 12 OWLs using ROBOT and fix ontology/prefix
So, the current set-up is just this:

./parse-all.sh

And everything gets parsed and converted to Turtle
automatically (I could add support for more output
serializations in the future, if needed).

The output turtles also seem to be opening correctly
by Protege, even though I just tested a few
(and the URI also appear to be correct).

Robot jar and sh come from this repo:

https://github.com/gbad-project/robot

As to the ontology/prefix issue, it was
originally described in an earlier commit:

42626795e1d3176c390002aad3d91ac5f40dd3a6

I reviewed the W3C recommendation on that,
which distinguishes between hash and slash
namespaces:

http://www.w3.org/2001/sw/BestPractices/VM/http-examples/2006-01-18/#naming

From this and also from exploring the original
drawio's from the examples dir, it became clear
that the -o option has to be just the base URI
("vocabulary URI") and -p has to be a slash URI.

What prompted me to revisit this was that
the ROBOT-mediated conversion from the original
set-up led to /schema/authority being set as both
@base and @prefix in the Turtle, which did not
appear correct.

Thank you to Daniel Garijo from stack overflow
for suggesting that W3C recommendation:

https://stackoverflow.com/a/33105358