Test attempt to use authority ontology directly in graph
I even uploaded it to GraphDB and visualized and
queries; but that seems to be a bad idea - that can
now be seen clearly. The reason is because these
additional deprecated entities are of no help
but are making querying more complex. Also, mapping
of literals appears to require hardcoding anyway.