parser v0.2.1-main.sha.1171107+gbad.8: allow absent x or y attrib
This gave an error (see parse-all.log) whereas in the drawing, the absent
attributes actually meant that the element touched an edge of the
background; thus, I commented these errors out, but not sure if any
malignant cases of absent x or y still exist.

Also, one more rico:accumulationDate was identified during parsing,
so fixing it here - see F 47

By the way, I am now using  ./parse-all.sh | tee parse-all.log
to ensure that the outputs from all the nested scripts are logged.

This patch allowed for the successful processing of all the 7 graphs
that weren't parsed previously - also included in the commit.